-Reapplied the new UI with a truetype font -Increased the number of games shown on one screen -Increased the number of characters shown per-title -Slowed down scrolling up/down and sped up left/right -Fixed Nintendont not running without an SD card inserted -Fixed a warning in Patch.c if DEBUG_PATCH was disabled
NOTE: Don't bother asking for a full GUI, coverflow, or anything else graphics related.  This is as good as it's going to get.  Use another loader if you want something nicer.  The purpose was to make the text easier to read on small screens, and to add flexibility for the devs by making font size customizable.